al hacer el T2 vi varias falencias que intente cubrir en este nuevo comit, ahora este si nos dara las bases para el T2
