+++ .gitignore
@@ -1,24 +1,23 @@
-# Compiled class file
+# Archivos compilados
 *.class
 
-# Log file
+# Archivos de logs
 *.log
 
-# BlueJ files
-*.ctxt
+# Configuración de IDEs (Eclipse, IntelliJ, VSCode)
+.vscode/
+.idea/
+*.iml
 
-# Mobile Tools for Java (J2ME)
-.mtj.tmp/
+# Directorios de build
+bin/
+out/
+build/
 
-# Package Files #
-*.jar
-*.war
-*.nar
-*.ear
-*.zip
-*.tar.gz
-*.rar
+# Archivos temporales del sistema operativo
+.DS_Store
+Thumbs.db
 
-# virtual machine crash logs, see http://www.java.com/en/download/help/error_hotspot.xml
-hs_err_pid*
-replay_pid*
+# Archivos de copia de seguridad
+*.bak
+*.tmp
