--- Dragon_Quest_VIII_T1/.gitignore
@@ -1,23 +1,0 @@
-# Archivos compilados
-*.class
-
-# Archivos de logs
-*.log
-
-# Configuración de IDEs (Eclipse, IntelliJ, VSCode)
-.vscode/
-.idea/
-*.iml
-
-# Directorios de build
-bin/
-out/
-build/
-
-# Archivos temporales del sistema operativo
-.DS_Store
-Thumbs.db
-
-# Archivos de copia de seguridad
-*.bak
-*.tmp
