+++ src/Enemigo.java
@@ -11,24 +11,28 @@ public class Enemigo extends Personaje {
 
     @Override
     public void atacar(Personaje enemigo) {
-        int danio = this.getAtaque() - enemigo.getDefensa();
-        if (danio < 0) {
-            danio = 0;
+        if (getEstado() != null && getEstado().getNombre().equals("Paralizado")) {
+            System.out.println(getNombre() + " está paralizado y no puede atacar.");
+            return;
         }
-
+        int danio = this.getAtaque() - enemigo.getDefensa();
+        if (danio < 0) danio = 0;
         enemigo.setVidaHp(enemigo.getVidaHp() - danio);
         if (enemigo.getVidaHp() <= 0) {
             enemigo.setVive(false);
             enemigo.setVidaHp(0);
             System.out.println(enemigo.getNombre() + " ha sido derrotado.");
         } else {
-            System.out.println(this.getNombre() + " ataca a " + enemigo.getNombre() + " causando " + danio + " puntos de dolor.");
+            System.out.println(this.getNombre() + " ataca a " + enemigo.getNombre() + " causando " + danio + " puntos de daño.");
         }
     }
 
     public void accionAutomatica(Personaje enemigo) {
         int decision = random.nextInt(100);
-
+        if (getEstado() != null && getEstado().getNombre().equals("Paralizado")) {
+            System.out.println(getNombre() + " está paralizado y no puede actuar.");
+            return;
+        }
         if (tipo.equalsIgnoreCase("agresivo")) {
             atacar(enemigo);
         } else if (tipo.equalsIgnoreCase("defensivo") && decision < 30) {
@@ -39,17 +43,13 @@ public class Enemigo extends Personaje {
     }
 
     public void defender() {
-        System.out.println(this.getNombre() + " se prepara para defenderse del proximo ataque.");
+        System.out.println(this.getNombre() + " se prepara para defenderse del próximo ataque.");
         setDefensa(getDefensa() + 10);
     }
 
     public String getTipo() {
-        return tipo;
-    }
-
-    public void setTipo(String tipo) {
-        this.tipo = tipo;
-    }
+         return tipo; 
+        }
+    public void setTipo(String tipo) { this.tipo = tipo; }
 }
-
